+++ myqueue/local.py
@@ -67,6 +67,7 @@ class LocalScheduler(Scheduler):
 
 
 class Server:
+    """Run tasks in subprocesses."""
     def __init__(self,
                  config: Configuration,
                  cores: int = 1,
@@ -84,21 +85,33 @@ class Server:
         self.aiotasks: Dict[int, asyncio.Task] = {}
         self.folder = self.config.home / '.myqueue'
 
-    async def main(self) -> None:
-        self.server = await asyncio.start_server(
-            self.recv, '127.0.0.1', self.port)
+    def start(self) -> None:
+        """Start server and wait for cammands."""
+        try:
+            asyncio.run(self._start())
+        except asyncio.exceptions.CancelledError:
+            pass
+
+    async def _start(self) -> None:
+        for _ in range(5):
+            try:
+                self.server = await asyncio.start_server(
+                    self.recv, '127.0.0.1', self.port)
+            except OSError:
+                self.port -= 1
+            else:
+                break
+        else:  # no break
+            raise OSError('Could not find unused port!')
+
+        print('Port:', self.port)
 
         async with self.server:  # type: ignore
             await self.server.serve_forever()
             await self.server.wait_closed()
 
-    def start(self) -> None:
-        try:
-            asyncio.run(self.main())
-        except asyncio.exceptions.CancelledError:
-            pass
-
     async def recv(self, reader: Any, writer: Any) -> None:
+        """Recieve command from client."""
         data = await reader.read(4096)
         cmd, *args = pickle.loads(data)
         print('COMMAND:', cmd, *args)
@@ -220,4 +233,4 @@ class Server:
 
 
 if __name__ == '__main__':
-    asyncio.run(Server(Configuration.read()).main())
+    asyncio.run(Server(Configuration.read())._start())